source/v2/requirements.txt
requests
python-decouple
firebase
firebase-admin
gitpython